improved debugging in cleanup

+++ pkg.sh
@@ -153,17 +153,23 @@ cleanup() {
     if [ "$cleanup" = 1 ]; then
         for arg in $BUILD_ORDER; do
             log_debug "In cleanup: Running cleanup"
+
             for repo in $repository_list; do
                 [ -d "$repo/$arg" ] && arg_dir="$repo/$arg"
             done
+
             change_directory "$arg_dir"
 
-            log_debug "In cleanup: rm -rf $(realpath "./$arg/build/")"
             # Tarballs, git repos, and patches were downloaded to build dir
-            [ -d ./build/ ] && rm -rf ./build/
+            if [ -d ./build/ ]; then
+                log_debug "In cleanup: rm -rf $(realpath "./$arg/build/")"
+                rm -rf ./build/
+            fi
 
-            log_debug "In cleanup: rm -rf $(realpath "./$arg/install/")"
-            [ -d ./install/ ] && rm -rf ./install/
+            if [ -d ./install/ ]; then
+                log_debug "In cleanup: rm -rf $(realpath "./$arg/install/")"
+                rm -rf ./install/
+            fi
 
             log_debug "In cleanup: cd $pwd"
             cd "$pwd" || true
